backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
google-generativeai>=0.3.1
python-dotenv>=1.0.0
pyodbc>=4.0.39
pandas>=2.0.0
tenacity>=8.2.2
python-multipart>=0.0.6
pydantic>=2.0.0
openpyxl>=3.1.2